// Font-end/mingalaba_management/src/app/_models/policy/enum/policy.type.enum.ts
export enum PolicyTypeEnum {
  _ = 'policy.policy.all',
  // _DISCOUNT_AMOUNT_ORDER = 'DISCOUNT_AMOUNT_ORDER',
  _DISCOUNT_PERCENT_ORDER_P1 = 'policy.title.DISCOUNT_PERCENT_ORDER_P1',
  _DISCOUNT_PERCENT_ORDER_P2 = 'policy.title.DISCOUNT_PERCENT_ORDER_P2',
  _REACHED_AMOUNT_DISCOUNT_PERCENT_ORDER_P1 = 'policy.title.REACHED_AMOUNT_DISCOUNT_PERCENT_ORDER_P1',
  _REACHED_AMOUNT_DISCOUNT_PERCENT_ORDER_P2 = 'policy.title.REACHED_AMOUNT_DISCOUNT_PERCENT_ORDER_P2',
  _INCENTIVE_RATE_AMOUNT_REVENUE = 'policy.title.INCENTIVE_RATE_AMOUNT_REVENUE',
  _INCENTIVE_RATE_PERCENT_REVENUE = 'policy.title.INCENTIVE_RATE_PERCENT_REVENUE',
  _LOYALTY_POINT_AMOUNT_ORDER = 'policy.title.LOYALTY_POINT_AMOUNT_ORDER',
  _LOYALTY_POINT_AMOUNT_LOGIN = 'policy.title.LOYALTY_POINT_AMOUNT_LOGIN',
  _LOYALTY_POINT_AMOUNT_REVIEW_PRODUCT = 'policy.title.LOYALTY_POINT_AMOUNT_REVIEW_PRODUCT',
  _LOYALTY_POINT_AMOUNT_NETWORK_EXPANSION_L1_L1_NEW_ACTIVATION = 'policy.title.LOYALTY_POINT_AMOUNT_NETWORK_EXPANSION_L1_L1_NEW_ACTIVATION',
  _LOYALTY_POINT_AMOUNT_NETWORK_EXPANSION_L1_L1_REAL_DEVELOPMENT = 'policy.title.LOYALTY_POINT_AMOUNT_NETWORK_EXPANSION_L1_L1_REAL_DEVELOPMENT',
  // _LOYALTY_POINT_AMOUNT_NETWORK_EXPANSION_L2_L2_NEW_ACTIVATION = 'LOYALTY_POINT_AMOUNT_NETWORK_EXPANSION_L2_L2_NEW_ACTIVATION',
  // _LOYALTY_POINT_AMOUNT_NETWORK_EXPANSION_L1_OPERATE_FOR_DAYS='LOYALTY_POINT_AMOUNT_NETWORK_EXPANSION_L1_OPERATE_FOR_DAYS',
  _LOYALTY_POINT_AMOUNT_NETWORK_EXPANSION_L1_PURCHASE_IN_DAYS = 'policy.title.LOYALTY_POINT_AMOUNT_NETWORK_EXPANSION_L1_PURCHASE_IN_DAYS',
  _INCENTIVE_RATE_AMOUNT_NETWORK_EXPANSION_L1_L1_NEW_ACTIVATION = 'policy.title.INCENTIVE_RATE_AMOUNT_NETWORK_EXPANSION_L1_L1_NEW_ACTIVATION',
  _INCENTIVE_RATE_AMOUNT_NETWORK_EXPANSION_L1_L1_REAL_DEVELOPMENT = 'policy.title.INCENTIVE_RATE_AMOUNT_NETWORK_EXPANSION_L1_L1_REAL_DEVELOPMENT',
  // _INCENTIVE_RATE_AMOUNT_NETWORK_EXPANSION_L2_L2_NEW_ACTIVATION = 'INCENTIVE_RATE_AMOUNT_NETWORK_EXPANSION_L2_L2_NEW_ACTIVATION',
  _INCENTIVE_RATE_INSTALL_APP_EXISTS_PURCHASE = 'policy.title.INCENTIVE_RATE_INSTALL_APP_EXISTS_PURCHASE',
  _LOYALTY_POINT_INSTALL_APP_EXISTS_PURCHASE = 'policy.title.LOYALTY_POINT_INSTALL_APP_EXISTS_PURCHASE'
  // _OTHER = 'policy.other'
}